llm/orchestrator-server/src/main/python/app/pyproject.toml
[tool.poetry]
name = "llm_orchestrator"
version = "0.1.0"
description = ""
authors = ["Mohamed ASSOUKTI <mohamed.assoukti@partnre.com>"]
readme = "README.md"
packages = [{include = "llm_orchestrator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.25.0"
gunicorn = "^21.2.0"
pydantic-settings="^2.1.0"
fastapi = "^0.108.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
tiktoken = "^0.5.2"
opensearch-py = "^2.4.2"
path = "^16.9.0"
colorlog = "^6.8.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
